update benchmarks to run sequentially, bench ws vs unix (#45)
