--- __tests__/fixtures/largePayload.json
@@ -1,33 +1,0 @@
-{
-  "data": [
-    {
-      "type": "articles",
-      "id": "1",
-      "attributes": {
-        "title": "Example article",
-        "body": "The shortest article. Ever.",
-        "created": "2015-05-22T14:56:29.000Z",
-        "updated": "2015-05-22T14:56:28.000Z"
-      },
-      "relationships": {
-        "author": {
-          "data": {
-            "id": "42",
-            "type": "people"
-          }
-        }
-      }
-    }
-  ],
-  "included": [
-    {
-      "type": "people",
-      "id": "42",
-      "attributes": {
-        "name": "John",
-        "age": 21,
-        "gender": "male"
-      }
-    }
-  ]
-}
