+++ package.json
@@ -63,12 +63,12 @@
   },
   "devDependencies": {
     "@types/ws": "^8.5.5",
-    "@vitest/ui": "^1.0.1",
+    "@vitest/ui": "^1.1.0",
     "prettier": "^3.0.0",
     "tempy": "^3.1.0",
     "tsup": "^7.2.0",
     "typescript": "^5.2.2",
-    "vitest": "^1.0.1"
+    "vitest": "^1.1.0"
   },
   "scripts": {
     "check": "tsc --noEmit && npx prettier . --check",
